Refactoring models classes into individual files
Based on https://github.com/jonfroehlich/makeabilitylabwebsite/issues/845
