+++ website/models/__init__.py
@@ -1,0 +1,15 @@
+from .banner import Banner
+from .keyword import Keyword
+from .news import News
+from .person import Person
+from .photo import Photo
+from .position import Position
+from .poster import Poster
+from .project import Project
+from .project_header import Project_header
+from .project_role import Project_Role
+from .project_umbrella import Project_umbrella
+from .publication import Publication
+from .sponsor import Sponsor
+from .talk import Talk
+from .video import Video
